UX: Redesign and refactor penalty modals
This merges the two modals code to remove duplication and implements
a more consistent design.
