+++ app/assets/javascripts/admin/addon/templates/components/admin-penalty-post-action.hbs
