+++ app/assets/javascripts/admin/addon/templates/components/admin-penalty-reason.hbs
@@ -1,0 +1,18 @@
+<div class="penalty-reason-controls">
+  {{#if (eq @penaltyType "suspend")}}
+    <label class="suspend-reason-title">{{i18n "admin.user.suspend_reason_title"}}</label>
+    <ComboBox @content={{this.reasons}} @value={{this.selectedReason}} @class="suspend-reason" @onChange={{action this.setSelectedReason}} />
+
+    {{#if this.isCustomReason}}
+      <TextField @value={{this.customReason}} @class="suspend-reason" @onChange={{action this.setCustomReason}} />
+    {{/if}}
+  {{else if (eq @penaltyType "silence")}}
+    <label class="silence-reason-title">{{html-safe (i18n "admin.user.silence_reason_label")}}</label>
+    <TextField @value={{this.customReason}} @class="silence-reason" @onChange={{action this.setCustomReason}} @placeholderKey="admin.user.silence_reason_placeholder" />
+  {{/if}}
+</div>
+
+<div class="penalty-message-controls">
+  <label>{{i18n "admin.user.suspend_message"}}</label>
+  <Textarea @value={{this.message}} class="suspend-message" placeholder={{i18n "admin.user.suspend_message_placeholder"}} />
+</div>
