+++ app/assets/javascripts/admin/addon/templates/components/silence-details.hbs
@@ -1,4 +1,4 @@
-<div class="reason-controls">
+<div class="penalty-reason-controls">
   <label>
     <div class="silence-reason-label">
       {{html-safe (i18n "admin.user.silence_reason_label")}}
