--- app/assets/javascripts/admin/addon/templates/components/suspension-details.hbs
@@ -1,25 +1,0 @@
-<div class="reason-controls">
-  <label>
-    <div class="suspend-reason-label">
-      {{#if this.siteSettings.hide_suspension_reasons}}
-        {{html-safe (i18n "admin.user.suspend_reason_hidden_label")}}
-      {{else}}
-        {{html-safe (i18n "admin.user.suspend_reason_label")}}
-      {{/if}}
-    </div>
-  </label>
-  <label>
-    {{i18n "admin.user.suspend_reason_title"}}
-  </label>
-  <ComboBox @content={{this.reasons}} @value={{this.selectedReason}} @class="suspend-reason" @onChange={{action this.setSelectedReason}} />
-  {{#if this.isCustomReason}}
-    <TextField @value={{this.customReason}} @class="suspend-reason" @onChange={{action this.setCustomReason}} />
-  {{/if}}
-</div>
-
-<label>
-  <div class="suspend-message-label">
-    {{i18n "admin.user.suspend_message"}}
-  </div>
-</label>
-<Textarea @value={{this.message}} class="suspend-message" placeholder={{i18n "admin.user.suspend_message_placeholder"}} />
