--- app/assets/javascripts/admin/addon/templates/modal/admin-silence-user.hbs
@@ -1,33 +1,0 @@
-<DModalBody @title="admin.user.silence_modal_title">
-  <ConditionalLoadingSpinner @condition={{this.loadingUser}}>
-
-    {{#if this.errorMessage}}
-      <div class="alert alert-error">{{this.errorMessage}}</div>
-    {{/if}}
-
-    <AdminPenaltyHistory @user={{this.user}} />
-
-    <div class="until-controls">
-      <label>
-        <FutureDateInput @class="silence-until" @label="admin.user.silence_duration" @clearable={{false}} @input={{this.silenceUntil}} @onChangeInput={{action (mut this.silenceUntil)}} />
-      </label>
-    </div>
-
-    <SilenceDetails @reason={{this.reason}} @message={{this.message}} />
-    {{#if this.postId}}
-      <PenaltyPostAction @postId={{this.postId}} @postAction={{this.postAction}} @postEdit={{this.postEdit}} />
-    {{/if}}
-
-    {{#if this.user.similar_users}}
-      <AdminPenaltySimilarUsers @type="silence" @user={{this.user}} @selectedUserIds={{this.otherUserIds}} />
-    {{/if}}
-
-  </ConditionalLoadingSpinner>
-
-</DModalBody>
-
-<div class="modal-footer">
-  <DButton @class="btn-danger perform-silence" @action={{action "silence"}} @disabled={{this.submitDisabled}} @icon="microphone-slash" @label="admin.user.silence" />
-  <DModalCancel @close={{route-action "closeModal"}} />
-  <ConditionalLoadingSpinner @condition={{this.loading}} @size="small" />
-</div>
