--- app/assets/javascripts/admin/addon/templates/modal/admin-suspend-user.hbs
@@ -1,39 +1,0 @@
-<DModalBody @title="admin.user.suspend_modal_title">
-  <ConditionalLoadingSpinner @condition={{this.loadingUser}}>
-
-    {{#if this.errorMessage}}
-      <div class="alert alert-error">{{this.errorMessage}}</div>
-    {{/if}}
-
-    {{#if this.user.canSuspend}}
-      <AdminPenaltyHistory @user={{this.user}} />
-
-      <div class="until-controls">
-        <label>
-          <FutureDateInput @class="suspend-until" @label="admin.user.suspend_duration" @clearable={{false}} @input={{this.suspendUntil}} @onChangeInput={{action (mut this.suspendUntil)}} />
-        </label>
-      </div>
-
-      <SuspensionDetails @reason={{this.reason}} @message={{this.message}} />
-      {{#if this.postId}}
-        <PenaltyPostAction @postId={{this.postId}} @postAction={{this.postAction}} @postEdit={{this.postEdit}} />
-      {{/if}}
-
-      {{#if this.user.similar_users}}
-        <AdminPenaltySimilarUsers @type="suspend" @user={{this.user}} @selectedUserIds={{this.otherUserIds}} />
-      {{/if}}
-    {{else}}
-      <div class="cant-suspend">
-        {{i18n "admin.user.cant_suspend"}}
-      </div>
-    {{/if}}
-
-  </ConditionalLoadingSpinner>
-
-</DModalBody>
-
-<div class="modal-footer">
-  <DButton @class="btn-danger perform-suspend" @action={{action "suspend"}} @disabled={{this.submitDisabled}} @icon="ban" @label="admin.user.suspend" />
-  <DModalCancel @close={{route-action "closeModal"}} />
-  <ConditionalLoadingSpinner @condition={{this.loading}} @size="small" />
-</div>
