+++ app/assets/javascripts/discourse/tests/acceptance/admin-penalize-user-test.js
@@ -1,3 +1,4 @@
+import { click, fillIn, visit } from "@ember/test-helpers";
 import {
   acceptance,
   count,
@@ -7,10 +8,9 @@ import {
   query,
   queryAll,
 } from "discourse/tests/helpers/qunit-helpers";
-import { click, fillIn, visit } from "@ember/test-helpers";
 import selectKit from "discourse/tests/helpers/select-kit-helper";
-import { test } from "qunit";
 import I18n from "I18n";
+import { test } from "qunit";
 
 acceptance("Admin - Suspend User", function (needs) {
   needs.user();
@@ -83,7 +83,7 @@ acceptance("Admin - Suspend User", function (needs) {
     await click(".suspend-user");
 
     assert.strictEqual(
-      count(".perform-suspend[disabled]"),
+      count(".perform-penalize[disabled]"),
       1,
       "disabled by default"
     );
@@ -94,9 +94,9 @@ acceptance("Admin - Suspend User", function (needs) {
     await fillIn("input.suspend-reason", "for breaking the rules");
     await fillIn(".suspend-message", "this is an email reason why");
 
-    assert.ok(!exists(".perform-suspend[disabled]"), "no longer disabled");
+    assert.ok(!exists(".perform-penalize[disabled]"), "no longer disabled");
 
-    await click(".perform-suspend");
+    await click(".perform-penalize");
 
     assert.ok(!exists(".suspend-user-modal:visible"));
     assert.ok(exists(".suspension-info"));
@@ -123,6 +123,50 @@ acceptance("Admin - Suspend User - timeframe choosing", function (needs) {
   test("shows correct timeframe options", async function (assert) {
     await visit("/admin/users/1234/regular");
     await click(".suspend-user");
+    await click(".future-date-input-selector-header");
+
+    const options = Array.from(
+      queryAll(`ul.select-kit-collection li span.name`).map((_, x) =>
+        x.innerText.trim()
+      )
+    );
+
+    const expected = [
+      I18n.t("time_shortcut.later_today"),
+      I18n.t("time_shortcut.tomorrow"),
+      I18n.t("time_shortcut.later_this_week"),
+      I18n.t("time_shortcut.start_of_next_business_week_alt"),
+      I18n.t("time_shortcut.two_weeks"),
+      I18n.t("time_shortcut.next_month"),
+      I18n.t("time_shortcut.two_months"),
+      I18n.t("time_shortcut.three_months"),
+      I18n.t("time_shortcut.four_months"),
+      I18n.t("time_shortcut.six_months"),
+      I18n.t("time_shortcut.one_year"),
+      I18n.t("time_shortcut.forever"),
+      I18n.t("time_shortcut.custom"),
+    ];
+
+    assert.deepEqual(options, expected, "options are correct");
+  });
+});
+
+acceptance("Admin - Silence User", function (needs) {
+  let clock = null;
+  needs.user();
+
+  needs.hooks.beforeEach(() => {
+    const timezone = loggedInUser().user_option.timezone;
+    clock = fakeTime("2100-05-03T08:00:00", timezone, true); // Monday morning
+  });
+
+  needs.hooks.afterEach(() => {
+    clock.restore();
+  });
+
+  test("shows correct timeframe options", async function (assert) {
+    await visit("/admin/users/1234/regular");
+    await click(".silence-user");
     await click(".future-date-input-selector-header");
 
     const options = Array.from(
