+++ app/assets/stylesheets/common/admin/admin_base.scss
@@ -1010,7 +1010,6 @@ a.inline-editable-field {
 @import "common/admin/settings";
 @import "common/admin/users";
 @import "common/admin/penalty";
-@import "common/admin/suspend";
 @import "common/admin/badges";
 @import "common/admin/emails";
 @import "common/admin/json_schema_editor";
